averge of three number

+++ function/qes1.java
@@ -3,19 +3,19 @@
 import java.util.*;
 public class qes1 {
 
-public static void avrgofThreeNUmb(int a,int b,int c){
-    int average  = (a+b+c)/3;
+public static void avrgofThreeNUmb(double a,double b,double c){
+    double average  = (a+b+c)/3;
     System.out.println("average of Three NUmber: " + average );
 }
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System .out.println("Enter first number ");
-        int a = sc.nextInt();
+        double a = sc.nextInt();
         System.out.println("Enter second number ");
-        int b = sc.nextInt();
+        double b = sc.nextDouble();
         System.out.println("Enter Third number");
-        int c = sc.nextInt();
+        double c = sc.nextDouble();
         avrgofThreeNUmb(a, b, c);
     }
 }
